backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
websockets==12.0
python-dotenv==1.0.1
pydantic==2.6.3
Pillow>=10.2.0
numpy>=1.26.4
python-multipart>=0.0.9
